requirements.txt
# requirements.txt
crewai
langchain-ollama  # The new, correct package
pydantic
python-dotenv